+++ ReactiveUI.Platforms/ReactiveUI.Android.csproj
@@ -67,9 +67,9 @@
     <Compile Include="Android\ActivityRoutedViewHost.cs" />
     <Compile Include="Android\AndroidCommandBinders.cs" />
     <Compile Include="Android\AndroidUIScheduler.cs" />
-    <Compile Include="Android\FlexibleCommandBinder.cs" />
     <Compile Include="Android\Properties\AssemblyInfo.cs" />
     <Compile Include="ComponentModelTypeConverter.cs" />
+    <Compile Include="FlexibleCommandBinder.cs" />
     <Compile Include="PlatformUnitTestDetector.cs" />
     <Compile Include="Registrations.cs" />
     <Compile Include="Android\PlatformOperations.cs" />
