+++ ReactiveUI.Platforms/ReactiveUI.Cocoa_Monotouch.csproj
@@ -71,6 +71,7 @@
     <Compile Include="Cocoa\RouterUINavigationController.cs" />
     <Compile Include="Cocoa\TargetActionCommandBinder.cs" />
     <Compile Include="Cocoa\ViewModelViewHost.cs" />
+    <Compile Include="FlexibleCommandBinder.cs" />
     <Compile Include="ComponentModelTypeConverter.cs" />
     <Compile Include="PlatformUnitTestDetector.cs" />
     <Compile Include="Registrations.cs" />
@@ -78,7 +79,6 @@
     <Compile Include="Cocoa\UIKitObservableForProperty.cs" />
     <Compile Include="Cocoa\UIKitObservableForPropertyBase.cs" />
     <Compile Include="Cocoa\Converters\DateTimeToNSDateConverter.cs" />
-    <Compile Include="Cocoa\FlexibleCommandBinder.cs" />
     <Compile Include="Cocoa\UIKitCommandBinders.cs" />
     <Compile Include="Cocoa\ReactiveNSView.cs" />
     <Compile Include="Cocoa\ReactiveNSViewController.cs" />
